pyproject.toml
[project]
name = "codereviewer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "dishka>=1.6.0",
    "fastapi[all]>=0.116.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-gigachat>=0.3.11",
    "langchain-huggingface>=0.3.1",
    "langchain-redis>=0.2.3",
    "langchain-weaviate>=0.0.5",
    "langgraph>=0.5.3",
    "langgraph-checkpoint>=2.1.0",
    "langgraph-checkpoint-redis>=0.0.8",
    "lark>=1.2.2",
    "llama-index>=0.12.52",
    "llama-index-embeddings-huggingface>=0.5.5",
    "llama-index-extractors-entity>=0.3.1",
    "llama-index-readers-file>=0.4.11",
    "llama-index-vector-stores-weaviate>=1.3.1",
    "loguru>=0.7.3",
    "markitdown[all]>=0.1.2",
    "mypy>=1.17.0",
    "networkx>=3.5",
    "pymupdf4llm>=0.0.27",
    "python-ulid>=3.0.0",
    "pytz>=2025.2",
    "redis>=6.2.0",
    "redisvl>=0.8.0",
    "spacy>=3.8.7",
    "span-marker>=1.7.0",
    "sqlalchemy>=2.0.41",
    "transformers>=4.53.2",
    "tree-sitter>=0.25.0",
    "tree-sitter-language-pack>=0.9.0",
    "weaviate-client>=4.16.1",
    "xmltodict>=0.14.2",
]
